[N-76] Use remote config values for player instead of editor config

+++ ZombierClient/Assets/Scripts/Data/WeaponData.cs
@@ -7,7 +7,7 @@ namespace Prototype.Data
     [CreateAssetMenu(fileName = "New Weapon Data", menuName = "Data/Weapon Data")]
     public class WeaponData : SerializedScriptableObject
     {
-        public int AttackRateRPM;
+        public float AttackRateRPM;
         public float AttackRange;
         public float Thrust = 3f;
         public float Recoil = 0.1f;
